requirements.txt
# OSNet ReID Requirements
# Python >= 3.8

# Deep Learning
torch>=1.10.0
torchvision>=0.11.0

# Configuration
PyYAML>=5.4.0

# Data Processing
numpy>=1.20.0
pandas>=1.3.0
Pillow>=8.0.0

# Training & Visualization
tqdm>=4.60.0
matplotlib>=3.3.0

# Evaluation
scikit-learn>=1.0.0

# ONNX Export (optional)
# onnx>=1.12.0
# onnxruntime>=1.12.0
